+++ CharacterOptionsPlus/CharacterOptionsPlus/Spells/BladedDash.cs
@@ -1,29 +1,28 @@
 ﻿using BlueprintCore.Actions.Builder;
+using BlueprintCore.Actions.Builder.ContextEx;
 using BlueprintCore.Blueprints.CustomConfigurators.UnitLogic.Abilities;
+using BlueprintCore.Blueprints.CustomConfigurators.UnitLogic.Buffs;
 using BlueprintCore.Blueprints.References;
 using BlueprintCore.Conditions.Builder;
+using BlueprintCore.Conditions.Builder.ContextEx;
 using BlueprintCore.Utils.Types;
 using CharacterOptionsPlus.Components;
 using CharacterOptionsPlus.Util;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Blueprints.JsonSystem;
-using Kingmaker.EntitySystem.Entities;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
-using Kingmaker.Enums.Damage;
-using Kingmaker.RuleSystem;
+using Kingmaker.PubSubSystem;
+using Kingmaker.RuleSystem.Rules;
+using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Abilities;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Abilities.Components;
-using Kingmaker.UnitLogic.Abilities.Components.Base;
+using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.Utility;
-using Owlcat.Runtime.Core.Utils;
 using System;
-using System.Collections.Generic;
-using UnityEngine;
 using static Kingmaker.UnitLogic.Commands.Base.UnitCommand;
 using static Kingmaker.Visual.Animation.Kingmaker.Actions.UnitAnimationActionCastSpell;
-using static TabletopTweaks.Core.MechanicsChanges.MetamagicExtention;
 using static UnityModManagerNet.UnityModManager.ModEntry;
 
 namespace CharacterOptionsPlus.Spells
@@ -35,8 +34,10 @@ namespace CharacterOptionsPlus.Spells
 
     internal const string DisplayName = "BladedDash.Name";
     private const string Description = "BladedDash.Description";
-    internal const string DisplayNameGreater = "BladedDash.Greater.Name";
+    private const string DisplayNameGreater = "BladedDash.Greater.Name";
     private const string DescriptionGreater = "BladedDash.Greater.Description";
+
+    private const string BuffName = "BladedDash.Buff";
 
     private const string IconPrefix = "assets/icons/";
     private const string IconName = IconPrefix + "gloriousheat.png";
@@ -63,6 +64,7 @@ namespace CharacterOptionsPlus.Spells
     {
       Logger.Log($"Configuring {FeatureName} (disabled)");
 
+      BuffConfigurator.New(BuffName, Guids.BladedDashBuff).Configure();
       AbilityConfigurator.New(FeatureName, Guids.BladedDashSpell).Configure();
       AbilityConfigurator.New(FeatureNameGreater, Guids.BladedDashGreaterSpell).Configure();
     }
@@ -71,72 +73,81 @@ namespace CharacterOptionsPlus.Spells
     {
       Logger.Log($"Configuring {FeatureName}");
 
-      AbilityConfigurator.NewSpell(FeatureName, Guids.BladedDashSpell, SpellSchool.Transmutation, canSpecialize: true)
+      var buff = BuffConfigurator.New(BuffName, Guids.BladedDashBuff)
+        .SetFlags(BlueprintBuff.Flags.HiddenInUi)
+        .AddNotDispelable()
+        .AddComponent<BladedDashBuff>()
+        .Configure();
+
+      var bladedDash = AbilityConfigurator.NewSpell(
+          FeatureName, Guids.BladedDashSpell, SpellSchool.Transmutation, canSpecialize: false)
         .SetDisplayName(DisplayName)
         .SetDescription(Description)
         .SetIcon(IconName)
         .SetCustomRange(30.Feet())
         .AllowTargeting(enemies: true, friends: true)
-        .SetAnimation(CastAnimationStyle.EnchantWeapon)
+        .SetAnimation(CastAnimationStyle.Special)
         .SetActionType(CommandType.Standard)
         .SetType(AbilityType.Spell)
         .SetNeedEquipWeapons(true)
-        .SetAvailableMetamagic(Metamagic.CompletelyNormal, Metamagic.Quicken, Metamagic.Reach)
+        .SetAvailableMetamagic(Metamagic.CompletelyNormal, Metamagic.Quicken)
         .AddToSpellLists(level: 2, SpellList.Bard, SpellList.Magus)
-        //.AddAbilityEffectRunAction(
-        //  actions: ActionsBuilder.New()
-        //    .Conditional(
-        //      conditions: ConditionsBuilder.New().Add<WantsToDropWeapon>(),
-        //      ifTrue: ActionsBuilder.New()
-        //        .SavingThrow(
-        //          SavingThrowType.Reflex,
-        //          onResult: ActionsBuilder.New()
-        //            .ConditionalSaved(
-        //              succeed: ActionsBuilder.New().Add<Disarm>(),
-        //              failed: dealDamage)),
-        //      ifFalse: dealDamage))
-        .AddAbilityCasterHasWeaponWithRangeType(WeaponRangeType.Melee)
+        .AddComponent(new AbilityCasterHasWeaponSubcategory(WeaponSubCategory.Melee))
+        .AddAbilityEffectRunAction(
+          actions: ActionsBuilder.New()
+            .CastSpell(AbilityRefs.DimensionDoorCasterOnly.ToString())
+            .ApplyBuff(buff, ContextDuration.Fixed(1), toCaster: true)
+            .MeleeAttack()
+            .RemoveBuff(buff, toCaster: true))
+        .Configure();
+
+      AbilityConfigurator.NewSpell(
+          FeatureNameGreater, Guids.BladedDashGreaterSpell, SpellSchool.Transmutation, canSpecialize: false)
+        .CopyFrom(bladedDash)
+        .SetDisplayName(DisplayNameGreater)
+        .SetDescription(DescriptionGreater)
+        .SetIcon(IconNameGreater)
+        .SetRange(AbilityRange.Projectile)
+        .AddToSpellLists(level: 5, SpellList.Bard, SpellList.Magus)
+        .AddComponent(new AbilityCasterHasWeaponSubcategory(WeaponSubCategory.Melee))
+        .AddAbilityDeliverProjectile(
+          projectiles: new() { ProjectileRefs.Kinetic_AirBlastLine00.ToString() },
+          length: 30.Feet(),
+          type: AbilityProjectileType.Line,
+          needAttackRoll: true)
+        .AddAbilityEffectRunAction(
+          actions: ActionsBuilder.New()
+            .Conditional(
+              ConditionsBuilder.New().IsMainTarget(),
+              ifTrue: ActionsBuilder.New().CastSpell(AbilityRefs.DimensionDoorCasterOnly.ToString()))
+            .Conditional(
+              ConditionsBuilder.New().IsEnemy(),
+              ifTrue: ActionsBuilder.New()
+                .ApplyBuff(buff, ContextDuration.Fixed(1), toCaster: true)
+                .MeleeAttack()
+                .RemoveBuff(buff, toCaster: true)))
         .Configure();
     }
 
-    [TypeId("c84fdb32-bc4f-4205-afdb-5b5464aec020")]
-    private class AbilityBladedDash : AbilityDeliverProjectile
+    [TypeId("4784aa37-5464-41b1-923f-8eff1a932a6d")]
+    private class BladedDashBuff :
+      UnitFactComponentDelegate, IInitiatorRulebookHandler<RuleCalculateAttackBonusWithoutTarget>
     {
-      private readonly bool Greater = false;
-
-      public AbilityBladedDash(bool greater)
-      {
-        Greater = greater;
-        Type = AbilityProjectileType.Line;
-      }
-
-      public AbilityBladedDash()
-      {
-
-      }
-      // TODO: Eh fuck it. This is possible but a lot of work. Let's just do it like CotW.
-
-      public override bool WouldTargetUnit(AbilityData ability, Vector3 targetPos, UnitEntityData unit)
+      public void OnEventAboutToTrigger(RuleCalculateAttackBonusWithoutTarget evt)
       {
         try
         {
-          var caster = ability.Caster.Unit;
-          var normalized = (targetPos - caster.Position).normalized;
-          var launchPos = caster.EyePosition + normalized * caster.Corpulence;
-          var rangeInMeters = ability.Blueprint.GetRange(ability.HasMetamagic(Metamagic.Reach), ability).Meters;
-          var context = ability.CreateExecutionContext(new(targetPos));
-          if (Greater)
-            return WouldTargetUnitLine(context, caster, launchPos, normalized.To2D(), rangeInMeters);
-          else
-          {
-            return false;
-          }
+          var chaMod = (Owner.Stats.GetStat(StatType.Charisma) as ModifiableValueAttributeStat).Bonus;
+          var intMod = (Owner.Stats.GetStat(StatType.Intelligence) as ModifiableValueAttributeStat).Bonus;
+          evt.AddModifier(chaMod > intMod ? chaMod : intMod, Fact, ModifierDescriptor.Circumstance);
         }
         catch (Exception e)
         {
-          Logger.LogException("AbilityBladedDash.WouldTargetUnit", e);
+          Logger.LogException("BladedDashBuff.OnEventAboutToTrigger", e);
         }
       }
+
+      public void OnEventDidTrigger(RuleCalculateAttackBonusWithoutTarget evt) { }
     }
   }
 }