+++ CharacterOptionsPlus/changelog.md
@@ -3,6 +3,8 @@
 * New Spells:
     * Touch of Blindness - Includes support for multiple touch charges
     * Shadow Trap
+    * Bladed Dash and Bladed Dash, Greater
+    * 
 
 * Updates
     * Added Localized Duration to some spells